Move function to paragraph module
